Implement lazy git (#10)
* removed the lsp from packer and added lualine via lazygit

* removed telescope and added via lazy.nvim

* added onderk theme 🎭

* comments are back

* basic treesitter 🌳 added

* LSP/nvim-cmp snippets are now with lazy.nvim 💪

* closes #8 vimPlug replaced with Lazy.vim

The very first vim plugin manager I cam acrossed was vundle but it
was messy.

Luckicly there was vimplug. It served it purpose well but it was
sometime hard to setup. Becuase in tutorials people have used
packer.nvim and nicely you can initialize the plugin right there.

That was the big motivation but that I found Lazy.nvim that is picking
up the pace. It was so easy to move from vimplug to Lazy.nvim. I was
worried it would take month and mindset to transition. But I did it
three days. Or I can say 8 hours in total.

With this big change changed I used mason for managing lsp changes. That
I am also very happy with. I kind of ignore/skipped the UI to manage lsp
but more of writing configuration.

* added package json